Production

See merge request bikram.dhoju/olen-elib!1

+++ src/pustakalaya_apps/document/signals.py
@@ -5,8 +5,8 @@ from django.db import transaction
 from django.db.models.signals import post_save, pre_delete, m2m_changed
 from .models import Document, DocumentFileUpload
 from .tasks import convert_pdf
-
-
+from pustakalaya_apps.core.utils import send_mail_on_user_submission
+from django.contrib.auth.models import  User
 
 
 @receiver(m2m_changed, sender=Document.keywords.through)
@@ -19,6 +19,18 @@ from .tasks import convert_pdf
 @receiver(post_save, sender=Document)
 @transaction.atomic
 def index_or_update_document(sender, instance, **kwargs):
+    # send an email to user when the document is published.
+
+    # By pass for unpublished items
+    if instance.published == "no":
+        return
+
+    if instance.published == "yes":
+        # added to restrict the if user is super user
+        # if not sender is User.is_superuser:
+        print("whois sender=",sender,"is suerp =",User.is_superuser)
+        send_mail_on_user_submission(item=instance)
+
 
     if instance.license is not None:
         if instance.license:
@@ -31,6 +43,11 @@ def index_or_update_document(sender, instance, **kwargs):
 @receiver(pre_delete, sender=Document)
 @transaction.atomic
 def delete_document(sender, instance, **kwargs):
+
+     # By pass for unpublished items
+    if instance.published == "no":
+        return
+
     # Delete an index first before instance in db.
     instance.delete_index()
 
@@ -40,6 +57,8 @@ def pdfto_image(sender, instance, **kwargs):
     """
     Convert pdf to images
     """
+    # dissable this feature no use now
+    return
 
     # grab the file path
     file_full_path = instance.upload.path
