+++ src/pustakalaya_apps/video/signals.py
@@ -4,6 +4,7 @@
 from django.db.models.signals import post_save, pre_delete, m2m_changed
 from django.dispatch import receiver
 from django.db import transaction
+from pustakalaya_apps.core.utils import send_mail_on_user_submission
 
 from .models import Video
 
@@ -17,20 +18,35 @@ from .models import Video
 @transaction.atomic
 def index_or_update_video(sender, instance, **kwargs):
     """Update or create an instance to index server."""
+
+    # By pass for unpublished items
+    if instance.published == "no":
+        return 
+    
+
+    if instance.published == "yes":
+        send_mail_on_user_submission(item=instance)
+
     if instance.license is not None:
         if instance.license.license:
             instance.license_type = instance.license.license
 
     # TODO: use logging system
-    #print("Instance", sender, instance)
-
     instance.index()
+
+     
 
 
 @receiver(pre_delete, sender=Video)
 @transaction.atomic
 def delete_video(sender, instance, **kwargs):
     """Delete an item from index server."""
+
+      # By pass for unpublished items
+    if instance.published == "no":
+        return 
+
+        
     # TODO: Use logging system
     print("Delete index object")
     instance.delete_index()
